% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_bernoulli.R
\name{dbernalt}
\alias{dbernalt}
\alias{pbernalt}
\title{The Bernoulli Distribution}
\usage{
dbernalt(x, mu, log = FALSE)

pbernalt(q, mu, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{a numeric vector.}

\item{lower.tail}{logical; if TRUE (default),}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
dbernalt gives the density
}
\description{
The Bernoulli Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rbinom(10, 1, .25)
dbernalt(x, .25)

pbernalt(x, .25)

}
