% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{dbinomalt}
\alias{dbinomalt}
\title{The Binomial Distribution}
\usage{
dbinomalt(x, mu, log = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
dbinomalt gives the density
}
\description{
The Binomial Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rbinom(100, 10, .5)
dbinomalt(x, .5)

}
