% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{dgammaalt}
\alias{dgammaalt}
\alias{pgammaalt}
\alias{rgammaalt}
\title{The Inverse Gaussian Distribution}
\source{
Hardin, James William., and Joseph Hilbe. Generalized Linear Models and Extensions. Stata Press, 2012.
}
\usage{
dgammaalt(x, mu, sigma, log = FALSE)

pgammaalt(q, mu, sigma, lower.tail = TRUE, log.p = FALSE)

rgammaalt(n, mu, sigma)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{sigma}{a number indicating the standard deviation.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{a numeric vector.}

\item{lower.tail}{logical; if TRUE (default),}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. Must be length 1}
}
\value{
dgammaalt gives the density
}
\description{
The Inverse Gaussian Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rgammaalt(10, 3, 4)
dgammaalt(x, 3, 4)

pgammaalt(x, 3, 4)

}
