% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_geometric.R
\name{dgeomalt}
\alias{dgeomalt}
\alias{pgeomaltalt}
\title{The Geometric Distribution}
\usage{
dgeomalt(x, mu, log = FALSE)

pgeomaltalt(q, mu, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{a numeric vector.}

\item{lower.tail}{logical; if TRUE (default),}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
dgeomalt gives the density
}
\description{
The Geometric Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rgeom(10, .5)
dgeomalt(x, .5)

pgeomaltalt(x, .5)

}
