% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{dnbinomalt}
\alias{dnbinomalt}
\title{The Negative Binomial Distribution}
\source{
Hardin, James William., and Joseph Hilbe. Generalized Linear Models and Extensions. Stata Press, 2012.
}
\usage{
dnbinomalt(x, mu, size, log = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{size}{target for number of successful trials.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
dnbinomalt gives the density
}
\description{
The Negative Binomial Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rnbinom(10, 7, .5)
dnbinomalt(x, .5, 7)

}
