% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{dpoisalt}
\alias{dpoisalt}
\alias{ppoisalt}
\alias{rpoisalt}
\title{The Poisson Distribution}
\source{
Hardin, James William., and Joseph Hilbe. Generalized Linear Models and Extensions. Stata Press, 2012.
}
\usage{
dpoisalt(x, mu, log = FALSE)

ppoisalt(q, mu, lower.tail = TRUE, log.p = FALSE)

rpoisalt(n, mu)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{a numeric vector.}

\item{lower.tail}{logical; if TRUE (default),}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. Must be length 1}
}
\value{
dpoisalt gives the density
}
\description{
The Poisson Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rpoisalt(10, 1)

dpoisalt(x, 1)

ppoisalt(x, 1)

}
