% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{dbinomalt}
\alias{dbinomalt}
\title{The Binomial Distribution}
\source{
Hardin, James William., and Joseph Hilbe. Generalized Linear Models and Extensions. Stata Press, 2012.
}
\usage{
dbinomalt(x, mu, size, log = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{size}{number of trials (zero or more).}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
dbinomalt gives the density
}
\description{
The Binomial Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rbinom(10, 2, .5)
dbinomalt(x, .5, 2)

}
