% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{dgammaalt}
\alias{dgammaalt}
\alias{pgammaalt}
\title{The Inverse Gaussian Distribution}
\source{
Hardin, James William., and Joseph Hilbe. Generalized Linear Models and Extensions. Stata Press, 2012.
}
\usage{
dgammaalt(x, mu, sigma, log = FALSE)

pgammaalt(q, mu, sigma, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{sigma}{a number indicating the standard deviation.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{a numeric vector.}

\item{lower.tail}{logical; if TRUE (default),}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
dgammaalt gives the density
}
\description{
The Inverse Gaussian Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rgamma(n = 10, shape = 2, rate = 1)
dgammaalt(x, 1, 1)

pgammaalt(x, 1, 1)

}
