% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{dgeomalt}
\alias{dgeomalt}
\title{The Geometric Distribution}
\usage{
dgeomalt(x, mu, log = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
dgeomalt gives the density
}
\description{
The Geometric Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rgeom(100, .5)
dgeomalt(x, .5)

}
