% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{dnbinomalt}
\alias{dnbinomalt}
\alias{pnbinomalt}
\alias{rnbinomalt}
\title{The Negative Binomial Distribution}
\source{
Hardin, James William., and Joseph Hilbe. Generalized Linear Models and Extensions. Stata Press, 2012.
}
\usage{
dnbinomalt(x, mu, size, log = FALSE)

pnbinomalt(q, mu, size, lower.tail = TRUE, log.p = FALSE)

rnbinomalt(n, mu, size)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{size}{target for number of successful trials.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{a numeric vector.}

\item{lower.tail}{logical; if TRUE (default),}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. Must be length 1}
}
\value{
dnbinomalt gives the density
}
\description{
The Negative Binomial Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rnbinomalt(10, .5, 7)

dnbinomalt(x, .5, 7)

pnbinomalt(x, .5, 7)

}
