% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{dnormalt}
\alias{dnormalt}
\title{The Normal Distribution}
\usage{
dnormalt(x, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{sigma}{a number indicating the standard deviation.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
dnormalt gives the density
}
\description{
The Normal Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rnorm(100, 0, 1)
dnormalt(x, 0, 1)

}
