% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{dpoisalt}
\alias{dpoisalt}
\title{The Poisson Distribution}
\source{
Hardin, James William., and Joseph Hilbe. Generalized Linear Models and Extensions. Stata Press, 2012.
}
\usage{
dpoisalt(x, mu, log = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{mu}{a number indicating the mean.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
dpoisalt gives the density
}
\description{
The Poisson Distribution
}
\examples{
library(altForm)

set.seed(1)
x <- rpois(10, 1)
dpoisalt(x, 1)

}
